Move history model to store
