+++ client/src/components/History/Modals/CopyModal.vue
@@ -45,12 +45,11 @@
 </template>
 
 <script>
-import { History } from "components/History/model/History";
 import { mapActions, mapGetters } from "vuex";
 
 export default {
     props: {
-        history: { type: History, required: true },
+        history: { type: Object, required: true },
     },
     data() {
         return {
