+++ client/src/store/historyStore/collectionElementsStore.js
@@ -8,7 +8,7 @@
 import Vue from "vue";
 import { LastQueue } from "utils/promise-queue";
 import { urlData } from "utils/url";
-import { mergeArray } from "./utilities";
+import { mergeArray } from "./model/utilities";
 
 const limit = 100;
 const queue = new LastQueue();
