+++ client/src/store/historyStore/model/utilities.js
